refactoring: giving a new order to componenets acoording if they handle state or not
