+++ src/app/snippets/[snippets_tech]/page.tsx
@@ -1,6 +1,6 @@
-import { ReusableBanner } from "@/components/complex/reusable-banner/ReusableBanner";
-import { RecentSnippets } from "@/components/complex/recent-snippets/RecentSnippets";
-import { TitleComponent } from "@/components/basic/titles/TitleComponent";
+import { ReusableBanner } from "@/components/stateless/reusable-banner/ReusableBanner";
+import { RecentSnippets } from "@/components/stateless/recent-snippets/RecentSnippets";
+import { TitleComponent } from "@/components/stateless/titles/TitleComponent";
 import SnippetsByTech from "./SnippetsByTech";
 
 type Params = {
