+++ src/components/stateless/about-section/AboutSection.tsx
@@ -1,6 +1,4 @@
-"use client";
-
-import styles from "@/components/complex/about-section/about-section.module.scss";
+import styles from "./about-section.module.scss";
 import { AboutText } from "./AboutText";
 
 export const AboutSection = () => {
@@ -8,7 +6,7 @@ export const AboutSection = () => {
 		<>
 			<main className={styles.main_container}>
 				<div className={styles.main_about_glass}></div>
-			    <AboutText />
+				<AboutText />
 			</main>
 		</>
 	);
