+++ src/components/stateless/hint-section/HintSection.tsx
