+++ src/components/stateless/our-team/OurTeamCard.tsx
