+++ src/components/stateless/our-team/our_team.module.scss
