+++ src/components/stateless/slider/SliderCard.tsx
