+++ src/components/stateless/slider/slider.module.scss
